requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
cachetools
pydantic[email]
pyjwt